+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Admin Changes**
   - Users index is now paginated
+  - Admin dashboard now provides relevant reports and links
+  - Added a consent report by site and by month
 - **Gem Changes**
   - Updated to carrierwave 1.2.1
   - Updated to sitemap_generator 6.0.0